Conflicts:
	codemp/cgame/cg_local.h
	codemp/cgame/cg_predict.cpp
	codemp/cgame/cg_saga.c
	codemp/cgame/cg_view.cpp
	codemp/cgame/cg_weapons.c
	codemp/client/FxSystem.h
	codemp/client/cl_console.cpp
	codemp/client/cl_input.cpp
	codemp/client/cl_main.cpp
	codemp/client/snd_dma.cpp
	codemp/game/NPC_AI_Humanoid.cpp
	codemp/game/NPC_stats.cpp
	codemp/game/ai_main.c
	codemp/game/bg_panimate.cpp
	codemp/game/g_bot.c
	codemp/game/g_client.cpp
	codemp/game/g_cmds.c
	codemp/game/g_team.cpp
	codemp/game/w_saber.cpp
	codemp/ghoul2/G2_gore.h
	codemp/libpng/png.c
	codemp/libpng/png.h
	codemp/libpng/pngconf.h
	codemp/libpng/pngdebug.h
	codemp/libpng/pngerror.c
	codemp/libpng/pngget.c
	codemp/libpng/pnginfo.h
	codemp/libpng/pnglibconf.h
	codemp/libpng/pngmem.c
	codemp/libpng/pngpread.c
	codemp/libpng/pngpriv.h
	codemp/libpng/pngread.c
	codemp/libpng/pngrio.c
	codemp/libpng/pngrtran.c
	codemp/libpng/pngrutil.c
	codemp/libpng/pngset.c
	codemp/libpng/pngstruct.h
	codemp/libpng/pngtrans.c
	codemp/libpng/pngwio.c
	codemp/libpng/pngwrite.c
	codemp/libpng/pngwtran.c
	codemp/libpng/pngwutil.c
	codemp/libraries/jpeg-8c/README
	codemp/libraries/jpeg-8c/ioquake3-changes.diff
	codemp/libraries/jpeg-8c/jaricom.c
	codemp/libraries/jpeg-8c/jcapimin.c
	codemp/libraries/jpeg-8c/jcapistd.c
	codemp/libraries/jpeg-8c/jcarith.c
	codemp/libraries/jpeg-8c/jccoefct.c
	codemp/libraries/jpeg-8c/jccolor.c
	codemp/libraries/jpeg-8c/jcdctmgr.c
	codemp/libraries/jpeg-8c/jchuff.c
	codemp/libraries/jpeg-8c/jcinit.c
	codemp/libraries/jpeg-8c/jcmainct.c
	codemp/libraries/jpeg-8c/jcmarker.c
	codemp/libraries/jpeg-8c/jcmaster.c
	codemp/libraries/jpeg-8c/jcomapi.c
	codemp/libraries/jpeg-8c/jconfig.h
	codemp/libraries/jpeg-8c/jcparam.c
	codemp/libraries/jpeg-8c/jcprepct.c
	codemp/libraries/jpeg-8c/jcsample.c
	codemp/libraries/jpeg-8c/jctrans.c
	codemp/libraries/jpeg-8c/jdapimin.c
	codemp/libraries/jpeg-8c/jdapistd.c
	codemp/libraries/jpeg-8c/jdarith.c
	codemp/libraries/jpeg-8c/jdatadst.c
	codemp/libraries/jpeg-8c/jdatasrc.c
	codemp/libraries/jpeg-8c/jdcoefct.c
	codemp/libraries/jpeg-8c/jdcolor.c
	codemp/libraries/jpeg-8c/jdct.h
	codemp/libraries/jpeg-8c/jddctmgr.c
	codemp/libraries/jpeg-8c/jdhuff.c
	codemp/libraries/jpeg-8c/jdinput.c
	codemp/libraries/jpeg-8c/jdmainct.c
	codemp/libraries/jpeg-8c/jdmarker.c
	codemp/libraries/jpeg-8c/jdmaster.c
	codemp/libraries/jpeg-8c/jdmerge.c
	codemp/libraries/jpeg-8c/jdpostct.c
	codemp/libraries/jpeg-8c/jdsample.c
	codemp/libraries/jpeg-8c/jdtrans.c
	codemp/libraries/jpeg-8c/jerror.c
	codemp/libraries/jpeg-8c/jerror.h
	codemp/libraries/jpeg-8c/jfdctflt.c
	codemp/libraries/jpeg-8c/jfdctfst.c
	codemp/libraries/jpeg-8c/jfdctint.c
	codemp/libraries/jpeg-8c/jidctflt.c
	codemp/libraries/jpeg-8c/jidctfst.c
	codemp/libraries/jpeg-8c/jidctint.c
	codemp/libraries/jpeg-8c/jinclude.h
	codemp/libraries/jpeg-8c/jmemmgr.c
	codemp/libraries/jpeg-8c/jmemnobs.c
	codemp/libraries/jpeg-8c/jmemsys.h
	codemp/libraries/jpeg-8c/jmorecfg.h
	codemp/libraries/jpeg-8c/jpegint.h
	codemp/libraries/jpeg-8c/jpeglib.h
	codemp/libraries/jpeg-8c/jquant1.c
	codemp/libraries/jpeg-8c/jquant2.c
	codemp/libraries/jpeg-8c/jutils.c
	codemp/libraries/jpeg-8c/jversion.h
	codemp/libraries/libpng/png.c
	codemp/libraries/libpng/png.h
	codemp/libraries/libpng/pngconf.h
	codemp/libraries/libpng/pngdebug.h
	codemp/libraries/libpng/pngerror.c
	codemp/libraries/libpng/pngget.c
	codemp/libraries/libpng/pnginfo.h
	codemp/libraries/libpng/pnglibconf.h
	codemp/libraries/libpng/pngmem.c
	codemp/libraries/libpng/pngpread.c
	codemp/libraries/libpng/pngpriv.h
	codemp/libraries/libpng/pngread.c
	codemp/libraries/libpng/pngrio.c
	codemp/libraries/libpng/pngrtran.c
	codemp/libraries/libpng/pngrutil.c
	codemp/libraries/libpng/pngset.c
	codemp/libraries/libpng/pngstruct.h
	codemp/libraries/libpng/pngtrans.c
	codemp/libraries/libpng/pngwio.c
	codemp/libraries/libpng/pngwrite.c
	codemp/libraries/libpng/pngwtran.c
	codemp/libraries/libpng/pngwutil.c
	codemp/libraries/openal/ALut.lib
	codemp/libraries/openal/OpenAL32.lib
	codemp/libraries/png/rpng.cpp
	codemp/libraries/png/rpng.h
	codemp/libs/win32/zlib.lib
	codemp/openjk.vcxproj
	codemp/openjk.vcxproj.filters
	codemp/openjkded.vcxproj
	codemp/qcommon/cm_terrainmap.cpp
	codemp/qcommon/q_shared.h
	codemp/rd-common/tr_public.h
	codemp/rd-dedicated/tr_image.cpp
	codemp/rd-dedicated/tr_init.cpp
	codemp/rd-vanilla/rd-vanilla.vcxproj
	codemp/rd-vanilla/rd-vanilla.vcxproj.filters
	codemp/server/sv_game.cpp
	codemp/ui/ui_local.h
	codemp/ui/ui_shared.h
	codemp/zlib/zlib.lib
	lib/libpng/png.c
	lib/libpng/png.h
	lib/libpng/pngconf.h
	lib/libpng/pngdebug.h
	lib/libpng/pngerror.c
	lib/libpng/pngget.c
	lib/libpng/pnginfo.h
	lib/libpng/pnglibconf.h
	lib/libpng/pngmem.c
	lib/libpng/pngpread.c
	lib/libpng/pngpriv.h
	lib/libpng/pngread.c
	lib/libpng/pngrio.c
	lib/libpng/pngrtran.c
	lib/libpng/pngrutil.c
	lib/libpng/pngset.c
	lib/libpng/pngstruct.h
	lib/libpng/pngtrans.c
	lib/libpng/pngwio.c
	lib/libpng/pngwrite.c
	lib/libpng/pngwtran.c
	lib/libpng/pngwutil.c
	lib/zlib.lib

+++ codemp/qcommon/vm_interpreted.cpp
@@ -373,7 +373,7 @@ nextInstruction2:
 #ifdef DEBUG_VM
 		if (com_vmdebug->integer > 1)
 		{
-			if ( (unsigned)programCounter > vm->codeLength ) {
+			if ( (unsigned)programCounter > (unsigned)vm->codeLength ) {
 				Com_Error( ERR_DROP, "VM pc out of range" );
 			}
 
@@ -474,7 +474,7 @@ nextInstruction2:
 
 				src = (int *)&image[ r0&dataMask ];
 				dest = (int *)&image[ r1&dataMask ];
-				if ( ( (int)src | (int)dest | count ) & 3 ) {
+				if ( ( (intptr_t)src | (intptr_t)dest | count ) & 3 ) {
 					Com_Error( ERR_DROP, "OP_BLOCK_COPY not dword aligned" );
 				}
 				count >>= 2;
